Added error catching to endpoints

+++ src/main/java/com/example/demo/controllers/TodoController.java
@@ -3,10 +3,13 @@ package com.example.demo.controllers;
 import com.example.demo.models.Todo;
 import com.example.demo.services.TodoService;
 import org.apache.coyote.Response;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static org.springframework.http.ResponseEntity.status;
 
 @RestController
 @RequestMapping("/todos")
@@ -33,7 +36,12 @@ public class TodoController {
     }
 
     @PostMapping
-    public ResponseEntity<Todo> save(@RequestBody Todo todo) {
+    public ResponseEntity<?> save(@RequestBody Todo todo) {
+        if (todo.getTask().length() > 255) {
+            return new ResponseEntity<>("Task name exceeds character maximum (255).", HttpStatus.BAD_REQUEST);
+        } else if (todo.getTask().isEmpty()) {
+            return new ResponseEntity<>("Please enter a valid task.", HttpStatus.BAD_REQUEST);
+        }
         return ResponseEntity.ok(todoService.save(todo));
     }
 
@@ -48,12 +56,30 @@ public class TodoController {
     }
 
     @DeleteMapping
-    public void delete(@RequestBody Todo todo) {
-        todoService.delete(todo);
+    public ResponseEntity<?> delete(@RequestBody Todo todo) {
+        if (todo.getId() == 0 || todoService.findById(todo.getId()) == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        try {
+            todoService.delete(todo);
+            return ResponseEntity.ok("Todo deleted successfully.");
+        } catch (Exception e) {
+            return new ResponseEntity<>("Todo could not be deleted, please try again.", HttpStatus.BAD_REQUEST);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable long id) {
-        todoService.deleteById(id);
+    public ResponseEntity<?> deleteById(@PathVariable long id) {
+        if (id == 0 || todoService.findById(id) == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        try {
+            todoService.deleteById(id);
+            return ResponseEntity.ok("Todo deleted successfully.");
+        } catch (Exception e) {
+            return new ResponseEntity<>("Todo could not be deleted, please try again.", HttpStatus.BAD_REQUEST);
+        }
     }
 }
